Update README.md
The data for this telco company was obtained from kaggle. The data contained 21 characteristic features for over 7000 customer was cleaned and visualized for insights. The data was split into train and test data. The churn data was balanced using SMOTE technique. The train data was trained with deep learning to predict whether a customer will churn or otherwise. The model gave an accuracy score of 81.